Rename to config repo operations to prevent future collisions with existing config repo
