+++ server/webapp/WEB-INF/applicationContext-global.xml
@@ -28,7 +28,7 @@
 
     <context:annotation-config/>
 
-    <context:component-scan base-package="com.thoughtworks.go.apiv1.configrepo"/>
+    <context:component-scan base-package="com.thoughtworks.go.apiv1.configrepooperations"/>
     <context:component-scan base-package="com.thoughtworks.go.api.spring"/>
     <context:component-scan base-package="com.thoughtworks.go.api.pluginimages"/>
     <context:component-scan base-package="com.thoughtworks.go.spark.spring"/>
