+++ settings.gradle
@@ -27,7 +27,7 @@ include ':api:api-artifact-store-config-v1'
 include ':api:api-backups-v1'
 include ':api:api-base'
 include ':api:api-build_cause-v1'
-include ':api:api-config-repos-v1'
+include ':api:api-config-repo-operations-v1'
 include ':api:api-current-user-v1'
 include ':api:api-dashboard-v2'
 include ':api:api-data-sharing-reporting-v2'
